requirements.txt
aniso8601==9.0.1
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
cymysql==0.9.18
facebook-sdk==3.1.0
fastlogging==0.9.6
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.29
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.8
uWSGI==2.0.20
Werkzeug==2.0.2
